Updated tests such that no exceptions are thrown when searching for records which do not exist. Started using index message processors and looping through the results instead of catching the exception and applying string comparisons on the exception message.

+++ src/test/java/org/mule/modules/drupal/automation/testcases/IndexUsersTestCases.java
@@ -56,7 +56,7 @@ public class IndexUsersTestCases extends DrupalTestParent {
 			for (Integer userId : userIds) {
 				boolean found = false;
 				for (User user : retrievedUsers) {
-					if (user.getUid() == userId) {
+					if (user.getUid().intValue() == userId) {
 						found = true;
 						break;
 					}
